+++ src/QueryMap/Contrib/Operator/ContainsOperator.php
@@ -64,7 +64,7 @@ class ContainsOperator extends Operator
         $name = $this->filter->getName();
         $alias = $this->filter->getAlias();
 
-        $paramName = $name.'#'.$this->getName();
+        $paramName = $alias.'#'.$name.'#'.$this->getName();
         
         /** @var \Doctrine\ORM\QueryBuilder $query */
         $query = $adapter->getQuery();