+++ src/QueryMap/Contrib/Operator/LessThanOrEqualOperator.php
@@ -49,7 +49,7 @@ class LessThanOrEqualOperator extends Operator
         $name = $this->filter->getName();
         $alias = $this->filter->getAlias();
 
-        $paramName = $name.'#'.$this->getName();
+        $paramName = $alias.'#'.$name.'#'.$this->getName();
 
         /** @var \Doctrine\ORM\QueryBuilder $query */
         $query = $adapter->getQuery();