+++ src/QueryMap/Contrib/Operator/NotEqualOperator.php
@@ -52,7 +52,7 @@ class NotEqualOperator extends Operator
         $name = $this->filter->getName();
         $alias = $this->filter->getAlias();
 
-        $paramName = $name.'#'.$this->getName();
+        $paramName = $alias.'#'.$name.'#'.$this->getName();
 
         /** @var \Doctrine\ORM\QueryBuilder $query */
         $query = $adapter->getQuery();